Update for 2019; more nodes and many more edges
Even though it's only February, stuff is taking off. This removes all of
the invisible left->right arrows that previously existed in the hopes
that GraphViz will manage well enough without them, which it seems to be
doing okay with. It also changes the dashed lines to solid gray ones on
the basis that a very short horizontal dashed line (xNetHack to
SpliceHack) was indistinguishable from a solid one.