Level generator code fixes and enhancements
The code here was written long before I am committing it now, back when
I was working on algorithms with the tubhack developer. So I don't
particularly remember very well what its purpose is.

Adds a couple small functions for choosing a random x coordinate and
random y coordinate on the map.

Fixes matching_adjacent: prior to this it appeared to be partially
implemented. Now it correctly returns a list of matching coordinates
rather than just a count of the number of matching coordinates, and
correctly calls the function passed to it to determine what a "match" is
rather than just checking ot against a list of valid types. Also
corrects some calls of count_adjacent that were using the old style.