Add python program for graphing ammo mulch chances
As a comment in the program states, its monte carlo approach of
crunching actual ammo mulch numbers over and over again is obsolete,
because I derived the exact formula a while ago. It remains a future
task to strip that logic out of this program and replace it with the
specific formula, at which point it will still graph the mulch chances.