Add jonadab's forest generator script
This may not be of that much use to me since I don't do perl but here it
is anyway. It generates a forest, usually with a river running through
it or a large lake or swampy area, that can be open all the way to the
edge (unlike mkmap, usually) and a guaranteed random-walk path of
non-trees between the stairs (which oddly tends to disrupt the river
making it look like a series of water bodies).